Step 5: fix the inevitable compilation errors
It is almost 100% certain that you will now have a non-compiling project. The volume of compilation errors will obviously depend on the size of your project and its usage of the Play APIs. Attempts have been made wherever possible in supporting HMRC libraries to mitigate the effects of this.

One example of this is the requirement for filters to now provide an Akka Materializer. The microservice-bootstrap library provides a MicroserviceFilterSupport mixin trait which does this for you.

Similarly, the deprecation of FakeApplication in Play 2.5 is misleading as, in reality, it just won't work as intended for a Play 2.5 application that uses any form of DI anywhere (which HMRC Play 2.5 requires due to use of metrics-play). Therefore, the `WithFakeApplication` test mixin trait has replaced FakeApplication with GuiceApplicationBuilder which should, with any luck, transparently resolve this problem for you.

Consequently, for an empty microservice created using init-service, the only compilation error that needed fixing was to make filters work by adding the `MicroserviceFilterSupport` trait.

If, like me, you are using scalatest+play, you should have remembered, unlike me, to update it with the other dependencies in step 3. `"org.scalatestplus" %% "play" % "1.2.0" % scope` has been superseded by `"org.scalatestplus.play" %% "scalatestplus-play" % "1.5.1"`.

By the time you complete this step, your project should build. It may even pass your tests, depending how thorough they are. However, it is certainly not yet fully-functional as there is one absolutely critical final step.